[✨feat] 마크업 컴포넌트화 #17
