+++ src/pages/ZipDetail/Header.jsx
@@ -1,0 +1,8 @@
+export default function Header() {
+  return (
+    <header>
+      <button type="button">뒤로 가기</button>
+      <h1>주차장 상세보기</h1>
+    </header>
+  );
+}
