+++ src/pages/ZipDetail/InfoZip.jsx
@@ -1,0 +1,26 @@
+export default function InfoZip() {
+  return (
+    <section>
+      <div>
+        <h2>마포 공영주차장</h2>
+        <button type="button">즐겨찾기 추가</button>
+      </div>
+      <div>
+        <span>공영</span>
+        <span>전기차 충전소</span>
+        <span>화장실</span>
+        <span>공휴일 무료</span>
+      </div>
+      <div>
+        <span>서울 서대문구 충정로 60 10층</span>
+        <button type="button">주소 복사하기</button>
+      </div>
+      <div>최초 10분 2,500원 / 추가 5분당 500원</div>
+      <div>
+        <button type="button">주차비 미리보기</button>
+        <button type="button">전화</button>
+        <button type="button">길안내</button>
+      </div>
+    </section>
+  );
+}
