+++ src/pages/ZipDetail/RecentUpdate.jsx
@@ -1,0 +1,7 @@
+export default function RecentUpdate() {
+  return (
+    <div>
+      <span>최근 업데이트 날짜 - 2023년 2월 21일</span>
+    </div>
+  );
+}
