+++ src/pages/ZipDetail/Review.jsx
@@ -1,0 +1,3 @@
+export default function Review() {
+  return <section>리뷰</section>;
+}
